pyproject.toml
[project]
name = "first-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.0",
    "fastapi-users[sqlalchemy]>=15.0.3",
    "imagekitio>=5.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "sqlalchemy[asyncio]>=2.0.0",  # already implied but make sure
    "aiomysql>=0.2.0",             # async MySQL driver
]

